refactor: Mudanças na estrutura de pasta do model.
