+++ src/controllers/login/Auth.js
@@ -2,13 +2,13 @@ import { Models } from '../../models/index.js';
 
 export const Auth = async (req, res) => {
 
-    const responseModelValidation = await Models.users.Validation({ type: 'auth', params: req.body });    
+    const responseModelValidation = await Models.validations.FielsValidation({ type: 'auth', params: req.body });    
 
     if(!responseModelValidation.status){
         return res.status(401).json({ status: "error", description: responseModelValidation?.ValidationErrors });
     }
 
-    const responseModelSelectUser = await Models.auth.SelectUser({ email: req.body.email });
+    const responseModelSelectUser = await Models.database.SelectUser({ email: req.body.email });
 
     if(!responseModelSelectUser.status){
         return res.status(401).json({ status: "error", description: responseModelSelectUser?.errorMessage }); 
@@ -16,13 +16,13 @@ export const Auth = async (req, res) => {
 
     const userId = responseModelSelectUser?.data?.user_id;
 
-    const responseModelConfirmPass = await Models.auth.ConfirmPass({ pass: responseModelSelectUser?.data?.user_pass, passConfirm: req.body.pass });
+    const responseModelConfirmPass = await Models.validations.ConfirmPass({ pass: responseModelSelectUser?.data?.user_pass, passConfirm: req.body.pass });
 
     if(!responseModelConfirmPass.status){
         return res.status(401).json({ status: "error", description: responseModelConfirmPass?.errorMessage });
     }
 
-    const responseModelCreateJwt = await Models.auth.CreateJwt({ userId: userId });
+    const responseModelCreateJwt = await Models.services.CreateJwt({ userId: userId });
 
     if(!responseModelCreateJwt.status){
         return res.status(401).json({ status: "error", description: responseModelCreateJwt?.errorMessage });
